Move newrelic to fdapm package
