+++ fdapm/doc.go
@@ -1,0 +1,2 @@
+// Package with some helpful tools to use APM with your app.
+package fdapm
